pyproject.toml
[tool.poetry]
name = "sdbot"
version = "0.1.0"
description = ""
authors = ["octonezd <me@octonezd.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.4"
aiohttp = "^3.8.5"
python-dotenv = "^1.0.0"
bleach = "^6.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
